Use actual python parser to parse python files
The filbert.js based python parser produced a very bad parse tree,
which required a very brittle interpreter to make any use of it.
It therefore broke often in unexpected ways.

As we have a server extension now, we can use the python "ast" module
to parse and unparse the code exactly as well as the actual python
parser can.