Click Link to Select, CTRL+Click to add a Point
The old click to add point functionality grew quite annoying very quickly. In particular if we want to modify the graph to add some component, using shift+click to select a link ended up also selecting the node, because that particular behavior is meant for selecting multiple things.

By moving the add-point-to-link functionality to ctrl+click, we free up the regular click to be just a select, exactly how one expects it to be.