Add RunParallelThread Component
The RunParallelThread component runs its body on an independent thread. As that allows parallel execution of the body, it needs to be copied so the independent execution threads don't walk all over each other. They are also executed with a copy of the context. For that reason changes to the context don't propagate out.