Add support for running in environments that have a running asyncio loop
Jupyter's ipykernel starts an asyncio loop in the same thread as it is executing the
xircuits workflow. When using AsyncComponent in that environment directly, we would
encounter an exception thrown by AsyncToSync telling us to just await the thing.

A simple workaround to this is to not use ipykernel based execution by selecting the
"Terminal Run" option. However, that requires the user to know about the issue and
understand the workaround.

This change instead allows the workflow to adapt to its environment.